Fix: Avoid focused buttons interference with enter
When operating the calculator through both keyboard and mouse, clicks left the
last clicked button focused by the browser, thus interfering with 'enter' key
inputs, used to trigger the 'equals' function.

Commit '00de513' tried to solve the issue of some keyboard inputs default
behavior's interference by getting rid of all default behavior, but that didn't
completely solved the issue and had some unintended consequences. Therefore it's
been changed in favor of an approach that only disables the 'tab' key, since it
it could still could lead to unexpected behavior and is unnecessary when
operating the app only through a keyboard.